Implement Product class in Main
In this commit, the Product class has been implemented in the Main class. The Product class represents a product that can be added to the shopping cart, and has two properties - name and price. The constructor initializes these properties with the values passed as arguments. This class will be used later in the application to create an array of Product objects that can be added to the shopping cart.